feat: add GithubStatsSkeleton component
